requirements.txt
gradio
huggingface_hub
requests
python-dotenv
